// electron/src/shared/types.ts
export type ItemType =
  | 'file'
  | 'folder'
  | 'image'
  | 'video'
  | 'audio'
  | 'pdf'
  | 'text'
  | 'markdown'
  | 'url'
  | 'note'
  | 'unknown'